requirements.txt
amqp==5.3.1
annotated-types==0.7.0
async-timeout==5.0.1
backports-datetime-fromisoformat==2.0.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
Flask==3.1.3
Flask-Mail==0.10.0
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
marshmallow==4.0.1
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
six==1.17.0
SQLAlchemy==2.0.47
sqlmodel==0.0.34
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
zipp==3.23.0
